refactor: split converter syntax modules
